+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Full-text search in CP name (#823)
 - Local representation of node assets (#814)
 - Retrieve files to download permissions and storage address from localrep (#800)
+- Add API endpoint to serve all performances of a given compute plan (#907)
 
 ### Changed
 - Make possible to start the backend-server without orchestrator connection available (#811)